refactor(lark): isolate goal topic inbox ownership
